Splitted ApiSystem for every request to be made.
